Name more tests after their packages
It seems like a good practice for tests to always be grouped under their package name, so that it’s easy to find the tests when they fail.  It’s confusing to have the tests in packages like “spacebars-tests”, “spacebars-compiler”, and “templating” be in various groups like “spacebars - templates” and “spacebars - templating”.

I didn’t go through and enforce this convention in all packages, just a few, but I think it would be good and not too hard to do the rest sometime.